EventLoop.CommonTypes

Description

All common types throughout the EventLoop module

Documentation

type Pos = (Float, Float)

Type to express a position on the screen. It uses the format (x,y). As Canvas is used in the example implementation, remember that the lefttop corner is (0,0) and the leftbottom corner is (0, height of screen).

type Dimension = (Float, Float)

Type to express the dimension of an element. It uses the format (w,h).

type Element = [Char]

The name of a graphical element. It is used for Mouse to express the name of which element is clicked on.