EventLoop.EventProcessor

Description

Low-level eventloop framework. Uses IOMessages to communicate. This module can be used to express your own InputEvents and OutputEvents. For instance, we have a keyboard and mouse InputEvent example included in this module's implementation. But it might be that you want to define your own InputEvent structure. In that case, you can define your own FromJSON class to express how the literal IOMessage should be parsed into your own InputEvent datatype. The same goes for OutputEvent of course. An IOMessage is a String literal that the client sends. The accompanied standard webpage uses JSON as a protocol to express these String literals.

Documentation

eventloop

Arguments

:: (a -> IOMessage -> ([IOMessage], a))
The eventhandler eh that maps incoming IOMessages to outgoing IOMessages. It uses a variable of type a to store information in between calls to the event handler.

-> a
The begin store. It should be changed in the eventhandler function.

-> IO ()

Low-level function to call when dealing with IOMessages directly. If you want to use the example implementation using the InputEvent and OutputEvent , you should use the start function.

type IOMessage = JSONMessage

Type of the message that is used to communicate with the IO device. In this example implementation a webbrowser is used and the messages use JSON encoding.

readRequest :: Connection -> StdOutMutex -> IO IOMessage

This functions reads a request from the Connection and parses it into an IOMessage.

sendResponse :: ConnectionSendMutex -> Connection -> StdOutMutex -> IOMessage -> IO ()

This function takes an IOMessage and parses it into a response. Then it sends the response to the Connection.