EventLoop.Input

Documentation

data InputEvent

The central InputEvent type.

Constructors

InMouse Mouse
A Mouse event.

InKeyboard Keyboard
A Keyboard event.

InSysMessage SystemMessageIn
A SystemMessage event.

data Keyboard

Datatype to express the different Keyboard events.

Constructors

KeyPress KeyboardButton

type KeyboardButton = [Char]

Type to express how a KeyboardButton is modelled.

data Mouse

Datatype to express the different Mouse events. The Pos expresses where on the screen the event happened. The Element expresses on which top element on screen the event happened. The Element value is the name of the GObject.

Constructors

MouseClick MouseButton Pos Element
Expresses a complete MouseClick consisting of a MouseUp and a MouseDown.

MouseUp MouseButton Pos Element
Expresses when a MouseButton moves upward.

MouseDown MouseButton Pos Element
Expresses when a MouseButton is pushed down.

data MouseButton

The MouseButton on the mouse.

Constructors

MLeft

MRight

MMiddle

data SystemMessageIn

The different possible SystemMessageIns.

Constructors

Setup
A request for the Setup. This should be generated when the connection to a client is made. The answer should be a CanvasSetup message.

Background
A request for the Background. This should be generated when the connection to a client is made. The answer could be the background of the graphical application.

Time
When a timer has been spawn, each 'tick' a Time is generated by the client to let the server know it is time.