EventLoop.Input.Mouse

Description

This module expresses how the Mouse IO device is modelled in the example implementation.

Documentation

data Mouse

Datatype to express the different Mouse events. The Pos expresses where on the screen the event happened. The Element expresses on which top element on screen the event happened. The Element value is the name of the GObject.

Constructors

MouseClick MouseButton Pos Element
Expresses a complete MouseClick consisting of a MouseUp and a MouseDown.

MouseUp MouseButton Pos Element
Expresses when a MouseButton moves upward.

MouseDown MouseButton Pos Element
Expresses when a MouseButton is pushed down.

data MouseButton

The MouseButton on the mouse.

Constructors

MLeft

MRight

MMiddle