EventLoop.Input.SystemMessage

Description

This module expresses how the incoming SystemMessages are modelled in the example implementation.

Documentation

data SystemMessageIn

The different possible SystemMessageIns.

Constructors

Setup
A request for the Setup. This should be generated when the connection to a client is made. The answer should be a CanvasSetup message.

Background
A request for the Background. This should be generated when the connection to a client is made. The answer could be the background of the graphical application.

Time
When a timer has been spawn, each 'tick' a Time is generated by the client to let the server know it is time.