EventLoop.Json

Description

This JSON library can parse JSON formatted Strings into JSONMessages.

Documentation

data JSONMember

A row within a JSONObject consisting of a field name ('[Char]'/'String') and field content (JSONMessage).

Constructors

JSONMember [Char] JSONMessage

data JSONMessage

A JSONMessage.

Constructors

JSONFloat Float
A JSONFloat

JSONString [Char]
A JSONString

JSONBool Bool
A JSONBool

JSONObject [JSONMember]
A JSONObject. Consists of JSONMembers which express the rows within the object.

JSONArray [JSONMessage]
A JSONArray. Consists of JSONMessages which expresses the different elements in the array.

retrieve :: [Char] -> [JSONMember] -> Maybe JSONMessage

Retrieves the field content (JSONMessage) associated with the field name ('[Char]') from a list of JSONMembers.

retrieveError :: [Char] -> [JSONMember] -> JSONMessage

Same as retrieve but instead of returning a Maybe, it raises an error when the field name could not be found. Usable in situations where you know for sure the field name is in the list of JSONMembers.

class FromJSON a where

Class expressing that type a is parsable from a JSONMessage.

Methods

fromJsonMessage :: JSONMessage -> a

Function to call to parse a JSONMessage to type a.

class JSONAble a where

Class expressing that type a can be parsed into a JSONMessage.

Methods

toJsonMessage :: a -> JSONMessage

stringToJsonObject :: [Char] -> JSONMessage

Function to parse a JSONMessage from a String.