EventLoop.Main

Description

High-level starting point for the eventloop. Uses the example implementation of InputEvent and OutputEvent to model the keyboard, mouse and webbrowser. Assumes that the server connects to webbrowser which sends information in a JSON formatted String(IOMessage).

Documentation

start

Arguments

:: (a -> InputEvent -> ([OutputEvent], a))
Eventhandler from InputEvent to a list of OutputEvent. Also takes a variable of type a that holds the current state of information. Returns a changed variable of type a that will be called with next handler call.

-> a
The begin state of variable a.

-> IO ()

High-level function to start an eventloop. The eventloop takes an InputEvent and outputs an OutputEvent. Starting place for the example implementation of the eventloop. It takes a variable of type a to hold information in between handler calls.