EventLoop.Output.Single

Description

Makeshift server to output a single OutputEvent to the client. The function outSingle is the heart of this module.

Documentation

outSingle :: OutputEvent -> IO ()

Outputs a single OutputEvent. Right now only Draw events are implemented. The server automatically determines the maximum Dimensions of the picture and sends a Setup containing those Dimensions.