EventLoop.Output.SystemMessage

Description

This module expresses how the outgoing SystemMessages are modelled in the example implementation.

Documentation

data SystemMessageOut

The different possible SystemMessageOuts.

Constructors

CanvasSetup Dimension
Answer to the Setup containing the dimensions of the canvas that will be used.

Timer TimeData
Request to create a timer at the clientside that will generate a Time message each 'tick'.

Close
A request for the client to completely close the connection to the server.

data TimeData

Constructors

On Int
Tells that the timer should be on with the time in ms

Off
Tells that the timer should be off