FPPrac.Graphics.NormalizeNumber

Description

In the EventLoop module, Floats are used to express positions and other characteristics. As that package is also used in the functional programming lab at the University of Twente, compatibility with the FPPrac module is needed. That module expresses a Number type which abstracts from Int and Double types. This module is used to normalize both the Number type and the original Num class instances defined in Haskell.

Documentation

class Num a => NormalizeNumber a where

Class to express that the instance a is able to be normalized to a Float.

Methods

normalize :: a -> Float

The function that can be used to normalize a to a Float.